2.0.17

+++ Prism/Scripts/PrismUtils/MediaProducts.py
@@ -211,6 +211,9 @@ class MediaProducts(object):
 
     @err_catcher(name=__name__)
     def getVersionsFromIdentifier(self, identifier, locations=None):
+        if not identifier:
+            return
+
         locationData = self.core.paths.getRenderProductBasePaths()
         searchLocations = []
         for locData in locationData:
@@ -242,7 +245,7 @@ class MediaProducts(object):
 
     @err_catcher(name=__name__)
     def getVersionStackContextFromPath(self, filepath, mediaType=None):
-        context = self.core.paths.getRenderProductData(filepath)
+        context = self.core.paths.getRenderProductData(filepath, mediaType=mediaType)
 
         if mediaType:
             context["mediaType"] = mediaType
@@ -269,6 +272,32 @@ class MediaProducts(object):
         return versionData
 
     @err_catcher(name=__name__)
+    def getVersion(self, entity, identifier, mediaType=None, version=None):
+        mediaType = mediaType or "3drenders"
+        version = version or "latest"
+        idf = entity.copy()
+        idf["identifier"] = identifier
+        idf["mediaType"] = mediaType
+        if version == "latest":
+            versionData = self.getLatestVersionFromIdentifier(idf)
+        else:
+            versions = self.getVersionsFromIdentifier(idf)
+            versionData = None
+            for ver in versions:
+                if ver["version"] == version:
+                    versionData = ver
+
+        return versionData
+
+    @err_catcher(name=__name__)
+    def getFileFromVersion(self, version, aov=None):
+        if aov:
+            version["aov"] = aov
+
+        file = self.getFilePatternFromVersion(version)
+        return file        
+
+    @err_catcher(name=__name__)
     def getVersionsFromContext(self, context, keys=None, locations=None):
         locationData = self.core.paths.getRenderProductBasePaths()
         searchLocations = []
@@ -293,9 +322,9 @@ class MediaProducts(object):
                 versionData += self.core.projects.getMatchingPaths(template)
 
             for data in versionData:
-                c = copy.deepcopy(context)
+                c = self.getDeepCopy(context)
                 c.update(data)
-                if self.core.products.getIntVersionFromVersionName(c["version"]) is None and c["version"] != "master":
+                if self.core.products.getIntVersionFromVersionName(c["version"]) is None and c["version"] != "master" and os.getenv("PRISM_SHOW_INVALID_VERSION_NAMES", "0") == "0":
                     continue
 
                 c["paths"] = [data.get("path")]
@@ -311,6 +340,32 @@ class MediaProducts(object):
                     continue
 
         return versions
+
+    @err_catcher(name=__name__)
+    def isPicklable(self, value):
+        import pickle
+        try:
+            pickle.dumps(value)
+            return True
+        except (pickle.PicklingError, TypeError):
+            return False
+
+    @err_catcher(name=__name__)
+    def getDeepCopy(self, context):
+        if not isinstance(context, dict):
+            return context
+
+        try:
+            newDict = copy.deepcopy(context)
+        except:
+            newDict = {}
+            for key, value in context.items():
+                if self.isPicklable(value):
+                    newDict[key] = self.getDeepCopy(value)
+                else:
+                    print(f"Warning: Ignoring unpicklable value for key '{key}'")
+
+        return newDict
 
     @err_catcher(name=__name__)
     def getAovPathFromVersion(self, version):
@@ -425,6 +480,7 @@ class MediaProducts(object):
                 globPath = os.path.join(folder, context["source"].replace("#", "?"))
                 files = glob.glob(globPath)
             else:
+                files = []
                 for root, folders, files in os.walk(folder):
                     break
 
@@ -437,6 +493,7 @@ class MediaProducts(object):
                         if ext:
                             filepaths.append(rpath)
                         else:
+                            rdfiles = []
                             for rdroot, rdfolders, rdfiles in os.walk(rpath):
                                 break
 
@@ -612,6 +669,7 @@ class MediaProducts(object):
     ):
         framePadding = framePadding or ""
         comment = comment or ""
+        location = location or "global"
 
         versionUser = user or self.core.user
         basePath = self.core.paths.getRenderProductBasePaths()[location]
@@ -710,7 +768,7 @@ class MediaProducts(object):
         if not getExisting and not self.core.separateOutputVersionStack:
             fileName = self.core.getCurrentFileName()
             fnameData = self.core.getScenefileData(fileName)
-            if fnameData.get("type") in ["asset", "shot"]:
+            if fnameData.get("type") in ["asset", "shot"] and "version" in fnameData:
                 hVersion = fnameData["version"]
             else:
                 hVersion = self.core.versionFormat % self.core.lowestVersion
@@ -851,7 +909,8 @@ class MediaProducts(object):
         key = "renderVersions"
 
         location = self.getLocationFromPath(versionFolder)
-        context["project_path"] = self.core.paths.getRenderProductBasePaths()[location]
+        if location:
+            context["project_path"] = self.core.paths.getRenderProductBasePaths()[location]
 
         if "type" in context and "entityType" not in context:
             context["entityType"] = context["type"]
@@ -939,14 +998,26 @@ class MediaProducts(object):
     def getLocationFromPath(self, path):
         locDict = self.core.paths.getRenderProductBasePaths()
         nPath = os.path.normpath(path)
+        validLocs = []
         for location in locDict:
             if nPath.startswith(locDict[location]):
-                return location
+                validLocs.append(location)
+
+        if not validLocs:
+            return
+
+        validLocs = sorted(validLocs, key=lambda x: len(locDict[x]), reverse=True)
+        return validLocs[0]
 
     @err_catcher(name=__name__)
-    def getVersionPathFromMediaFilePath(self, path, mediaType):
-        entityType = self.core.paths.getEntityTypeFromPath(path)
+    def getVersionPathFromMediaFilePath(self, path, mediaType, entityType=None):
+        if not entityType:
+            entityType = self.core.paths.getEntityTypeFromPath(path)
+            if not entityType:
+                context = self.core.paths.getMediaProductData(path, mediaType=mediaType)
+                entityType = context.get("type")
 
+        key = None
         context = {"mediaType": mediaType}
         if mediaType == "playblasts":
             versionKey = "playblastVersions"
@@ -960,6 +1031,9 @@ class MediaProducts(object):
                 key = "renderFilesAssets"
             elif entityType == "shot":
                 key = "renderFilesShots"
+
+        if not key:
+            return
 
         location = self.getLocationFromPath(path)
         context["project_path"] = self.core.paths.getRenderProductBasePaths()[location]
@@ -1035,9 +1109,9 @@ class MediaProducts(object):
         masterDrive = os.path.splitdrive(masterPath)[0]
         drive = os.path.splitdrive(path)[0]
 
-        masterBase = self.getVersionPathFromMediaFilePath(masterPath, mediaType=context.get("mediaType"))
+        masterBase = self.getVersionPathFromMediaFilePath(masterPath, mediaType=context.get("mediaType"), entityType=context.get("type"))
         if isFilepath:
-            originBase = self.getVersionPathFromMediaFilePath(path, mediaType=context.get("mediaType"))
+            originBase = self.getVersionPathFromMediaFilePath(path, mediaType=context.get("mediaType"), entityType=context.get("type"))
         else:
             originBase = path
 
@@ -1197,7 +1271,7 @@ class MediaProducts(object):
             vpath = path
 
         logger.debug("removing master render version: %s" % vpath)
-        if os.path.exists(vpath):
+        if vpath and os.path.exists(vpath):
             try:
                 shutil.rmtree(vpath)
             except Exception as e:
@@ -1372,6 +1446,12 @@ class MediaProducts(object):
         self.ingestedFiles = []
         self.ingestCanceled = False
         self.ingestThreads = []
+        startFrame = 1
+        if entity.get("type") == "shot":
+            shotRange = self.core.entities.getShotRange(entity)
+            if shotRange:
+                startFrame = shotRange[0]
+
         with self.copyMsg as copyMsg:
             for idx, file in enumerate(files):
                 if self.ingestCanceled:
@@ -1379,7 +1459,7 @@ class MediaProducts(object):
 
                 kwargs["extension"] = os.path.splitext(file)[1]
                 if len(files) > 1:
-                    kwargs["framePadding"] = ("%%0%sd" % self.core.framePadding) % (idx + 1)
+                    kwargs["framePadding"] = ("%%0%sd" % self.core.framePadding) % (idx + startFrame)
 
                 if kwargs.get("mediaType") == "playblasts":
                     pbkwargs = kwargs.copy()
@@ -1506,3 +1586,23 @@ class MediaProducts(object):
                         outdatedVersions.append({"master": None, "latest": latestVersion})
 
         return outdatedVersions
+
+    @err_catcher(name=__name__)
+    def getGroupFromIdentifier(self, identifier):
+        identifierPath = self.getIdentifierPathFromEntity(identifier)
+        cfgPath = os.path.join(identifierPath, "identifiers" + self.core.configs.getProjectExtension())
+        group = self.core.getConfig(identifier.get("displayName"), "group", configPath=cfgPath)
+        return group
+
+    @err_catcher(name=__name__)
+    def setIdentifiersGroup(self, identifiers, group, projectWide=False):
+        identifierPath = self.getIdentifierPathFromEntity(identifiers[0])
+        cfgPath = os.path.join(identifierPath, "identifiers" + self.core.configs.getProjectExtension())
+        data = self.core.getConfig(configPath=cfgPath) or {}
+        for identifier in identifiers:
+            if identifier.get("displayName") not in data:
+                data[identifier.get("displayName")] = {}
+
+            data[identifier.get("displayName")]["group"] = group
+
+        self.core.setConfig(data=data, configPath=cfgPath)
